Some Fixes to restore offline stores, but not work correctly yet..

+++ trunk/datapack/sql/character_offline_trade.sql
@@ -9,7 +9,7 @@ CREATE TABLE `character_offline_trade` (
 CREATE TABLE `character_offline_trade_items` (
   `charId` int(10) NOT NULL DEFAULT '0',
   `item` int(10) NOT NULL DEFAULT '0',
-  `count` bigint(20) NOT NULL DEFAULT '0',
-  `price` bigint(20) NOT NULL DEFAULT '0',
+  `count` int(20) NOT NULL DEFAULT '0',
+  `price` int(20) NOT NULL DEFAULT '0',
   PRIMARY KEY (`charId`,`item`)
 ) ;
